\name{in.support.NoisyAR}
\alias{in.support.NoisyAR}
\title{Checks if parameters are valid for a Noisy AR model}
\usage{
  in.support.NoisyAR(nlss)
}
\arguments{
  \item{nlss}{Instance of the NoisyAR class}
}
\value{
  TRUE if the parameters are valid, FALSE otherwise
}
\description{
  Checks if parameters are valid for a Noisy AR model
}

