\name{loclike.logpdf}
\alias{loclike.logpdf}
\title{Log-pdf of the local likelihood of a Non linear state space model}
\usage{
  loclike.logpdf(nlss, ...)
}
\arguments{
  \item{nlss}{Instance of a nlss class}

  \item{...}{Passed to the specific implementation}
}
\description{
  Log-pdf of the local likelihood of a Non linear state
  space model
}
\seealso{
  \link{NoisyAR}
}

