\name{nltsa-package}
\alias{nltsa-package}
\alias{nltsa}
\docType{package}
\title{Nonlinear Time Series Analysis}
\description{
Contains data sets and scripts for Nonlinear Time Series, Theory, Methods,
    and Applications by Douc, Moulines, Stoffer }
\details{
\tabular{ll}{
Package: \tab nltsa\cr
Type: \tab Package\cr
Version: \tab 1.01\cr
Date: \tab 2016-11-27\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
LazyData: \tab no\cr
}
}
\references{See the webpage for the text: \url{http://www.stat.pitt.edu/stoffer/nltsa/} 
}
\keyword{ package }
\keyword{ ts }
