\name{Stovol}
\alias{Stovol}
\title{Create a Stochastic Volatility  model}
\usage{
  Stovol(phi, sigma, beta)
}
\arguments{
  \item{phi}{State auto-regressive parameter}

  \item{sigma}{State noise standard deviation}

  \item{beta}{Basis standard volatility}
}
\value{
  S3 object of class NoisyAR
}
\description{
  Create a Stochastic Volatility model
}

